+++ include/open62541pp/Server.h
@@ -7,16 +7,14 @@
 #include <vector>
 
 #include "open62541pp/Logger.h"
+#include "open62541pp/types/NodeId.h"
 
 // forward declaration open62541
 struct UA_Server;
-struct UA_ServerConfig;
 
 namespace opcua {
 
 // forward declaration
-class NodeId;
-
 template <typename ServerOrClient>
 class Node;
 
@@ -29,8 +27,8 @@ struct Login {
 /**
  * High-level server class.
  *
- * Exposes the most common functionality. Use the handle() and getConfig() method to get access
- * to the underlying UA_Server and UA_ServerConfig instances and use the full power of open6254.
+ * Exposes the most common functionality. Use the handle() method to get access the underlying
+ * UA_Server instance and use the full power of open6254.
  */
 class Server {
 public:
@@ -79,9 +77,6 @@ public:
 
     UA_Server* handle() noexcept;
     const UA_Server* handle() const noexcept;
-
-    UA_ServerConfig* getConfig() noexcept;
-    const UA_ServerConfig* getConfig() const noexcept;
 
 private:
     class Connection;